Move project TOML to common
